feat: add router for shuttle
