+++ src/pages/Shuttle/TokenList.js
@@ -1,0 +1,5 @@
+function TokenList() {
+  return <div>TokenList</div>
+}
+
+export default TokenList
